+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
   # core
   "core/tauri",
+  "core/tauri-runtime",
   "core/tauri-macros",
   "core/tauri-utils",
   "core/tauri-build",
