+++ core/tauri-runtime/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "tauri-runtime"
+version = "0.0.0"
+authors = [ "Tauri Programme within The Commons Conservancy" ]
+categories = [ "gui", "web-programming" ]
+license = "Apache-2.0 OR MIT"
+homepage = "https://tauri.studio"
+repository = "https://github.com/tauri-apps/tauri"
+description = "Runtime for Tauri applications"
+edition = "2018"
+
+[dependencies]
+serde = { version = "1.0", features = [ "derive" ] }
+serde_json = "1.0"
+thiserror = "1.0"
+uuid = { version = "0.8.2", features = [ "v4" ] }
+tauri-utils = { version = "1.0.0-beta-rc.1", path = "../tauri-utils" }
+wry = { git = "https://github.com/tauri-apps/wry", rev = "6bc97aff525644b83a3a00537316c46d7afb985b" }
+image = "0.23"
